improve performance when deployed under mod_perl2 environment.
append, instead of prepending, app-specific lib to @INC can help
reducing the number of error-ed stat in the system call analysis.

for the record. In one of my app, the number of syscalls reduced to
437 from 506 (per-request) (reduced ~ 15%) just by this @INC change.
And the number of requests served per-second is almost doubled out of
this change. real big win.

fyi, the syscall is analyzed with dstruss on mac os leopard.